requirements.txt
httpx>=0.25.0
asyncio>=3.4.3
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
tqdm>=4.65.0
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
rich>=13.4.0
loguru>=0.7.0
joblib>=1.3.0
scikit-learn>=1.3.0
tiktoken>=0.5.0
tabulate>=0.9.0
click>=8.1.0